requirements.txt
openai>=1.55.3
litellm>=1.51.0
python-dotenv==1.0.1
colorama>=0.4.6
python-telegram-bot>=21.9
discord.py>=2.4.0
duckduckgo-search==6.3.3
httpx==0.27.2
beautifulsoup4==4.12.3
cloverlabs-camoufox
setuptools<82
mem0ai>=1.0.11
mcp>=1.27.0
pyautogui>=0.9.54
pillow>=10.0.0
pytesseract>=0.3.10
opencv-python-headless>=4.8.0
pygetwindow>=0.0.9
textual>=0.86.2
